1. Merged logger into misc.
