+++ iris-mpc-utils/src/lib.rs
@@ -3,8 +3,6 @@ pub mod convertor;
 pub mod genesis;
 pub mod graphs;
 pub mod irises;
+pub mod misc;
 pub mod state;
 pub mod types;
-pub mod utils;
-
-pub use utils::fsys;
