--- iris-mpc-utils/src/utils/misc.rs
@@ -1,10 +1,0 @@
-/// Returns a boxed iterator over the first `limit` elements of `iter`.
-pub fn limited_iterator<I>(iter: I, limit: Option<usize>) -> Box<dyn Iterator<Item = I::Item>>
-where
-    I: Iterator + 'static,
-{
-    match limit {
-        Some(num) => Box::new(iter.take(num)),
-        None => Box::new(iter),
-    }
-}
