--- iris-mpc-utils/src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod fsys;
-pub(crate) mod logger;
-pub(crate) mod misc;
